+++ src/commons-only-untested.spec.ts
@@ -1,0 +1,8 @@
+// import untested service for code-coverage
+import {StringUtils} from "./commons/utils/string.utils";
+
+for (const a in [
+    StringUtils
+]) {
+    console.log('import unused modules for codecoverage');
+}