+++ src/commons-only-untested.spec.ts
@@ -14,3 +14,9 @@ for (const a in [
 ]) {
     console.log('import unused modules for codecoverage');
 }
+
+describe('Dummy-Test', () => {
+    it('should be true', () => {
+        expect(true).toBeTruthy();
+    });
+});