+++ docs/CHANGELOG.md
@@ -39,6 +39,7 @@
 - common: added optional responsetype to GenericSearchHttpAdapter for export  
 - common: disabled js-data-cache in generic-datastore because of objects with different loadDetailsProfiles loaded
 - common: fixed replacement of parameters in generic-sql-adapter - replace global
+- common: fixed solr-query-builder
  
 ### breaking changes
 - none